corrected main.py
