--- Workshop9_homework/end_menu.py
@@ -1,5 +1,0 @@
-import logg
-
-def output(result):
-    print(result)
-    logg.logging.debug("Program has finished")
